+++ backend/settings/settings.go
@@ -47,6 +47,11 @@ func ApplyUserDefaults(u *users.User) {
 	u.Perm = Config.UserDefaults.Permissions
 	u.ShowHidden = Config.UserDefaults.ShowHidden
 	u.DateFormat = Config.UserDefaults.DateFormat
+	u.DisableOnlyOfficeExt = Config.UserDefaults.DisableOnlyOfficeExt
+	u.ThemeColor = Config.UserDefaults.ThemeColor
+	u.GallerySize = Config.UserDefaults.GallerySize
+	u.QuickDownload = Config.UserDefaults.QuickDownload
+	u.LockPassword = Config.UserDefaults.LockPassword
 	if len(u.Scopes) == 0 {
 		u.Scopes = []users.SourceScope{
 			{