+++ backend/http/middleware.go
@@ -60,7 +60,9 @@ func withHashFileHelper(fn handleFunc) handleFunc {
 			}
 		}
 		data.path = strings.TrimSuffix(link.Path, "/") + "/" + strings.TrimPrefix(path, "/")
-
+		if path == "" || path == "/" {
+			data.path = link.Path
+		}
 		// Get file information with options
 		file, err := FileInfoFasterFunc(iteminfo.FileOptions{
 			Path:   data.path,
@@ -70,7 +72,7 @@ func withHashFileHelper(fn handleFunc) handleFunc {
 		})
 		file.Token = link.Token
 		if err != nil {
-			logger.Error(fmt.Sprintf("error fetching file info for share %v %v %v", hash, path, err))
+			logger.Error(fmt.Sprintf("error fetching file info for share. hash=%v path=%v error=%v", hash, data.path, err))
 			return errToStatus(err), fmt.Errorf("error fetching share from server")
 		}
 		// Set the file info in the `data` object
@@ -118,13 +120,16 @@ func withUserHelper(fn handleFunc) handleFunc {
 					if err != nil {
 						return http.StatusInternalServerError, err
 					}
-					newUser := storage.CreateUser(users.User{
+					err = storage.CreateUser(users.User{
 						Username: proxyUser,
 						NonAdminEditable: users.NonAdminEditable{
 							Password: hashpass, // hashed password that can't actually be used
 						},
 					}, false)
-					data.user, err = store.Users.Get(newUser)
+					if err != nil {
+						return http.StatusInternalServerError, err
+					}
+					data.user, err = store.Users.Get(proxyUser)
 					if err != nil {
 						return http.StatusInternalServerError, err
 					}
@@ -143,7 +148,6 @@ func withUserHelper(fn handleFunc) handleFunc {
 			return http.StatusUnauthorized, err
 		}
 		data.token = tokenString
-
 		var tk users.AuthToken
 		token, err := jwt.ParseWithClaims(tokenString, &tk, keyFunc)
 		if err != nil {