+++ CHANGELOG.md
@@ -4,15 +4,17 @@ All notable changes to this project will be documented in this file. For commit 
 
 ## v0.6.6-beta
 
-
+ **New Feature**:
+ - limit tar size creation to limit server burden. For example, don't let customers try to download the entire filesystem as a zip. see `server.maxArchiveSize` on config wiki.
 
  **Notes**:
  - disableUsedPercentage also hides text and source bar.
  - share errors show up in logs in more verbose way.
+ - archive creation occurs on disk rather than in memory, use `server.cacheDir` to determine where temp files are stored.
 
  **BugFixes**:
  - fix proxy user creation issue
-
+ - externalUrl prefix issue fixed for shares.
 
 ## v0.6.5-beta