+++ CHANGELOG.md
@@ -13,8 +13,10 @@ All notable changes to this project will be documented in this file. For commit 
  - archive creation occurs on disk rather than in memory, use `server.cacheDir` to determine where temp files are stored.
 
  **BugFixes**:
- - fix proxy user creation issue
+ - fix proxy user creation issue https://github.com/gtsteffaniak/filebrowser/issues/478
  - externalUrl prefix issue fixed for shares.
+ - fix File Opens Instead of Just Downloading https://github.com/gtsteffaniak/filebrowser/issues/480
+ - fix Download file name https://github.com/gtsteffaniak/filebrowser/issues/481
 
 ## v0.6.5-beta