+++ CHANGELOG.md
@@ -11,6 +11,9 @@ All notable changes to this project will be documented in this file. For commit 
  - disableUsedPercentage also hides text and source bar.
  - share errors show up in logs in more verbose way.
  - archive creation occurs on disk rather than in memory, use `server.cacheDir` to determine where temp files are stored.
+ - automatically ensures leading slash for scope
+   - https://github.com/gtsteffaniak/filebrowser/issues/472
+   - https://github.com/gtsteffaniak/filebrowser/issues/476
 
  **BugFixes**:
  - fix proxy user creation issue https://github.com/gtsteffaniak/filebrowser/issues/478