+++ CHANGELOG.md
@@ -17,7 +17,7 @@ All notable changes to this project will be documented in this file. For commit 
 
  **BugFixes**:
  - fix proxy user creation issue https://github.com/gtsteffaniak/filebrowser/issues/478
- - externalUrl prefix issue fixed for shares.
+ - externalUrl prefix issue fixed for shares. https://github.com/gtsteffaniak/filebrowser/issues/465
  - fix File Opens Instead of Just Downloading https://github.com/gtsteffaniak/filebrowser/issues/480
  - fix Download file name https://github.com/gtsteffaniak/filebrowser/issues/481